Pixel-Art-Generator/requirements.txt
opencv-python
numpy
streamlit
pillow
matplotlib